Move markdown functions to separate file
