+++ src/jsMain/kotlin/io/beatmaps/maps/mapinfo.kt
@@ -19,9 +19,9 @@ import io.beatmaps.index.ModalComponent
 import io.beatmaps.index.ModalData
 import io.beatmaps.index.links
 import io.beatmaps.playlist.addToPlaylist
+import io.beatmaps.util.textToContent
 import kotlinx.browser.window
 import kotlinx.html.InputType
-import kotlinx.html.Tag
 import kotlinx.html.id
 import kotlinx.html.js.onClickFunction
 import kotlinx.html.title
@@ -35,8 +35,6 @@ import react.RReadableRef
 import react.RState
 import react.ReactElement
 import react.createRef
-import react.dom.InnerHTML
-import react.dom.RDOMBuilder
 import react.dom.a
 import react.dom.button
 import react.dom.div
@@ -397,49 +395,8 @@ class MapInfo : RComponent<MapInfoProps, MapInfoState>() {
     }
 }
 
-fun String.transformURLIntoLinks() =
-    replace("\\b((https?|ftp)://)?[-a-zA-Z0-9@:%._+~#=]{2,256}\\.[A-Za-z]{2,6}\\b(/[-a-zA-Z0-9@:%_+.~#?&/=]*)*(?:/|\\b)".toRegex()) {
-        if (it.groupValues[1].isEmpty()) it.value else "<a target=\"_blank\" href=\"${it.value}\">${it.value}</a>"
-    }
-
-fun String.parseBoldMarkdown() =
-    replace("(^|\\s)(\\*\\*|__)(.*?)\\2".toRegex(RegexOption.MULTILINE)) {
-        "${it.groupValues[1]}<b>${it.groupValues[3]}</b>"
-    }
-
-fun String.parseItalicMarkdown() =
-    replace("(^|\\s)([*_])(.*?)\\2".toRegex(RegexOption.MULTILINE)) {
-        "${it.groupValues[1]}<i>${it.groupValues[3]}</i>"
-    }
-
-fun String.parseMapReference() =
-    replace("(^|\\s)#([\\da-f]+?)($|[^\\da-z])".toRegex(setOf(RegexOption.MULTILINE, RegexOption.IGNORE_CASE))) {
-        """${it.groupValues[1]}<a href="/maps/${it.groupValues[2].lowercase()}">#${it.groupValues[2]}</a>${it.groupValues[3]}"""
-    }
-
-fun String.parseUserReference() =
-    replace("(^|\\s)@([\\w.-]+?)($|[^\\w.-])".toRegex(setOf(RegexOption.MULTILINE, RegexOption.IGNORE_CASE))) {
-        """${it.groupValues[1]}<a href="/profile/username/${it.groupValues[2].lowercase()}">@${it.groupValues[2]}</a>${it.groupValues[3]}"""
-    }
-
 fun RBuilder.mapInfo(handler: MapInfoProps.() -> Unit): ReactElement {
     return child(MapInfo::class) {
         this.attrs(handler)
     }
-}
-
-// Kotlin IR be adding underscores everywhere
-class DangerousHtml(override var __html: String) : InnerHTML
-
-fun <T : Tag> RDOMBuilder<T>.textToContent(text: String) {
-    domProps.dangerouslySetInnerHTML = DangerousHtml(
-        text
-            .replace(Regex("<.+?>"), "")
-            .parseBoldMarkdown()
-            .parseItalicMarkdown()
-            .parseMapReference()
-            .parseUserReference()
-            .transformURLIntoLinks()
-            .replace("\n", "<br />")
-    )
 }
