+++ src/jsMain/kotlin/io/beatmaps/maps/testplay/feedback.kt
@@ -5,7 +5,7 @@ import external.TimeAgo
 import external.generateConfig
 import io.beatmaps.api.FeedbackUpdate
 import io.beatmaps.common.Config
-import io.beatmaps.maps.textToContent
+import io.beatmaps.util.textToContent
 import kotlinx.datetime.internal.JSJoda.Instant
 import kotlinx.html.TEXTAREA
 import kotlinx.html.js.onClickFunction
